+++ src/core/os.h
@@ -164,7 +164,7 @@ void os_thread_attach(void);
 void os_thread_detach(void);
 void os_thread_set_name(const char* name);
 
-void os_poll_events(void);
+void os_poll_events(double timeout);
 void os_on_quit(fn_quit* callback);
 void os_on_visible(fn_visible* callback);
 void os_on_focus(fn_focus* callback);