+++ src/core/os_android.c
@@ -290,10 +290,10 @@ void os_thread_set_name(const char* name) {
   //
 }
 
-void os_poll_events(void) {
+void os_poll_events(double timeout) {
   if (!state.app->destroyRequested) {
     struct android_poll_source* source;
-    int timeout = (state.app->window && state.app->activityState == APP_CMD_RESUME) ? 0 : -1;
+    int timeout = (state.app->window && state.app->activityState == APP_CMD_RESUME) ? (int) (timeout * 1000.) : -1;
     ALooper_pollOnce(timeout, NULL, NULL, (void**) &source);
 
     if (source) {